requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
Django==4.2.27
django-unfold==0.67.0
docopt==0.6.2
fonttools==4.60.2
num2words==0.5.14
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
sqlparse==0.5.5
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
